+++ spec/rails_app/config/routes.rb
@@ -31,6 +31,7 @@ AppRoot::Application.routes.draw do
     get :test_login_from_salesforce
     get :test_login_from_slack
     get :test_login_from_instagram
+    get :test_login_from_auth0
     get :login_at_test
     get :login_at_test_twitter
     get :login_at_test_facebook
@@ -45,6 +46,7 @@ AppRoot::Application.routes.draw do
     get :login_at_test_salesforce
     get :login_at_test_slack
     get :login_at_test_instagram
+    get :login_at_test_auth0
     get :test_return_to_with_external
     get :test_return_to_with_external_twitter
     get :test_return_to_with_external_facebook
@@ -59,6 +61,7 @@ AppRoot::Application.routes.draw do
     get :test_return_to_with_external_salesforce
     get :test_return_to_with_external_slack
     get :test_return_to_with_external_instagram
+    get :test_return_to_with_external_auth0
     get :test_http_basic_auth
     get :some_action_making_a_non_persisted_change_to_the_user
     post :test_login_with_remember